+++ contentScript.js
@@ -1,3 +1,5 @@
+"use strict";
+
 (() => {
   let manifest = chrome.runtime.getManifest();
   console.log(manifest.name + " v" + manifest.version);