+++ CHANGELOG.md
@@ -1,9 +1,13 @@
  # CHANGELOG
 
+ ### v0.1.5 (10 Aug 2023)
+ - send active tab screenshot
+ - "use strict"
+
  ### v0.1.4 (09 Aug 2023)
  - experimental: send video screenshot
- - TRY_FETCHING = false: always send dataURLby default
- - open in tab after active tab
+ - TRY_FETCHING = false: always send dataURL by default
+ - open Photopea after active tab
  - browser action: open Photopea
 
  ### v0.1.3